- yet another fix for file seeking

+++ HomerMultimedia/include/MediaSource.h
@@ -328,7 +328,7 @@ protected:
     AVCodecContext      *mCodecContext;
     int                 mMediaStreamIndex;
     double              mStartPtsUSecs /* only used in MediasourceNet and MediaSourceFile */, mSourceStartPts; // for synchronized playback we calculate the position within a media stream and write the value into PTS entry of an encoded packet
-    int64_t             mNumberOfFrames;
+    double              mNumberOfFrames;
     /* audio */
     AVFifoBuffer        *mRecorderSampleFifo;
     char                *mRecorderSamplesTempBuffer;
